+++ src/core/renderer/tiles/TilesRendererBase.js
@@ -1,7 +1,7 @@
 import { getUrlExtension } from '../utilities/urlExtension.js';
 import { LRUCache } from '../utilities/LRUCache.js';
 import { PriorityQueue } from '../utilities/PriorityQueue.js';
-import { runTraversal as optimizedRunTraversal } from './optimizedTraverseFunctions.js';
+import { runTraversal } from './traverseFunctions.js';
 import { UNLOADED, QUEUED, LOADING, PARSING, LOADED, FAILED } from '../constants.js';
 import { throttle } from '../utilities/throttle.js';
 import { traverseSet } from '../utilities/TraversalUtils.js';
@@ -844,7 +844,7 @@ export class TilesRendererBase {
 		this.prepareForTraversal();
 
 		// run traversal
-		optimizedRunTraversal( root, this );
+		runTraversal( root, this );
 
 		// remove any tiles that are loading but no longer used
 		this.removeUnusedPendingTiles();
