+++ src/core/renderer/tiles/traverseFunctions.js
